Create easy-stock.php
Easy setup, basic funcitonality that covers the most cases, setup includes 2 stocks, the secondary stock is callew with the custom field _secondary_stock. Feel free to exchange that with your own  stock / custom field to manager several stock locations in bulk without the use of bulky and to bloated plugins.